requirements.txt
xai_sdk
dotenv
typing
requests
flask
flask_cors
alpaca-py
